12-intro-to-langsmith/requirements.txt
langgraph
langgraph-sdk
langgraph-checkpoint-sqlite
langsmith>=0.2.0
langchain-community
langchain-core
langchain-openai
notebook
python-dotenv
lxml
scikit-learn
pandas
pyarrow